cleanup more transfer code
