+++ app/components/transfer/BaseTransferPage.scss
