--- app/components/transfer/TransferMasterKeyPage.scss
@@ -1,60 +1,0 @@
-@import '../../themes/mixins/loading-spinner';
-
-.component {
-  padding: 20px;
-
-  .button {
-    display: block !important;
-    margin: 20px auto 0;
-  }
-  
-  .body {
-    color: var(--theme-bordered-box-text-color);
-    font-family: var(--font-regular);
-    font-size: 14px;
-    line-height: 19px;
-  
-    .title {
-      line-height: 1.38;
-      margin-bottom: 4px;
-      font-size: 18px;
-      font-family: var(--font-semibold);
-    }
-  
-    .text {
-      font-size: 15px;
-      margin-bottom: 0.3%;
-      word-break: break-word;
-    }
-  
-    .instructionsList {
-      list-style-type: disc;
-      list-style-position: inside;
-      margin-bottom: 20px;
-    }
-  
-    .buttonsWrapper {
-      display: flex;
-      flex-direction: row-reverse;
-      justify-content: center;
-    }
-  
-    .submitWithPasswordButton {
-      &.spinning {
-        @include loading-spinner("../../assets/images/spinner-light.svg");
-      }
-    }
-  }
-}
-
-:global(.YoroiClassic) .component {
-  .button {
-    margin: 0 auto 0;
-  }
-}
-
-:global(.YoroiModern) .component {
-  fieldset {
-    background: var(--theme-settings-pane-background-color);
-  }
-}
