+++ app/components/widgets/forms/DaedalusMasterKeyInput.scss
@@ -1,0 +1,5 @@
+:global(.YoroiModern) {
+  fieldset {
+    background: var(--theme-settings-pane-background-color);
+  }
+}
