+++ app/containers/transfer/DaedalusTransferFormPage.js
@@ -2,8 +2,12 @@
 import React, { Component } from 'react';
 import { observer } from 'mobx-react';
 import { defineMessages, intlShape } from 'react-intl';
-import TransferMnemonicPage from '../../components/transfer/TransferMnemonicPage';
 import config from '../../config';
+import { join } from 'lodash';
+import { action, observable } from 'mobx';
+import ReactToolboxMobxForm from '../../utils/ReactToolboxMobxForm';
+import MnemonicInput from '../../components/widgets/forms/MnemonicInput';
+import BaseTransferPage from '../../components/transfer/BaseTransferPage';
 
 const messages = defineMessages({
   step0: {
@@ -32,30 +36,56 @@ export default class DaedalusTransferFormPage extends Component<Props> {
     intl: intlShape.isRequired
   };
 
+  static contextTypes = {
+    intl: intlShape.isRequired
+  };
+
+  @observable mnemonicForm: void | ReactToolboxMobxForm;
+
+  @action
+  setMnemonicFrom(form: ReactToolboxMobxForm) {
+    this.mnemonicForm = form;
+  }
+
+  submit = () => {
+    if (this.mnemonicForm == null) {
+      throw new Error('DaedalusTransferFormPage form not set');
+    }
+    this.mnemonicForm.submit({
+      onSuccess: (form) => {
+        const { recoveryPhrase } = form.values();
+        const payload = {
+          recoveryPhrase: join(recoveryPhrase, ' '),
+        };
+        this.props.onSubmit(payload);
+      },
+      onError: () => {}
+    });
+  };
+
   render() {
     const { intl } = this.context;
     const {
-      validWords,
-      onBack,
-      onSubmit,
-      mnemonicValidator,
       mnemonicLength,
-      classicTheme
     } = this.props;
     const message = mnemonicLength === config.wallets.DAEDALUS_PAPER_RECOVERY_PHRASE_WORD_COUNT
       ? messages.step0Paper
       : messages.step0;
-
     return (
-      <TransferMnemonicPage
-        onSubmit={onSubmit}
-        onBack={onBack}
-        mnemonicValidator={mnemonicValidator}
-        validWords={validWords}
+      <BaseTransferPage
+        onSubmit={this.submit}
+        onBack={this.props.onBack}
         step0={intl.formatMessage(message)}
-        mnemonicLength={mnemonicLength}
-        classicTheme={classicTheme}
-      />
+        classicTheme={this.props.classicTheme}
+      >
+        <MnemonicInput
+          setForm={(form) => this.setMnemonicFrom(form)}
+          mnemonicValidator={this.props.mnemonicValidator}
+          validWords={this.props.validWords}
+          mnemonicLength={this.props.mnemonicLength}
+          classicTheme={this.props.classicTheme}
+        />
+      </BaseTransferPage>
     );
   }
 }
